feat(bwanga): try to correct room
Sometimes wrong room is detected, and after the wrong room is finished, the correct room can't be entered as expected.
For example, room 6 is finished and room 7 is wrongly detected as room 4. After this "room 4" is finished, the correct room 7 can't be entered because moving from room 4 to room 7 is invalid.
We add counter for a "wrong room" which is continuously detected and try to correct current room to the "wrong room" when the counter exceeds threshold.

Signed-off-by: J.J. Martzki <mars14850@gmail.com>